requirements.txt
cycler==0.12.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
h5py==3.7.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.5.2
MinkowskiEngine==0.5.4
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
protobuf==3.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.12
tensorboardX==2.5.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
triton==2.3.0
typing-extensions==4.11.0
